portfolioApi/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
whitenoise==6.4.0
